# services/api-gateway/openapi.yaml
...
openapi: 3.1.0
info:
  title: RiskRadar API Gateway
  version: 1.0.0
  description: Public HTTP surface of the RiskRadar platform routed through the API Gateway.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
    description: Gateway root
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    X-Request-ID:
      in: header
      name: X-Request-ID
      required: false
      schema:
        type: string
      description: Correlation ID propagated by API Gateway
...
  schemas:
    ApiSuccess:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
      required: [success]
  responses:
    Error:
      description: Standardized error response
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
paths:
...
  /api/reports/createReport:
    post:
      summary: Create report
      tags: [reports]
      security: [ { bearerAuth: [] } ]
      operationId: postReportsCreateReport
      parameters:
        - $ref: '#/components/parameters/X-Request-ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, latitude, longitude, userId, reportCategory]
              properties:
                title: { type: string }
                description: { type: string }
                latitude: { type: number, format: double }
                longitude: { type: number, format: double }
                userId: { type: string, format: uuid }
                imageIds:
                  type: array
                  items: { type: string, format: uuid }
                reportCategory:
                  type: string
                  enum: [VANDALISM, INFRASTRUCTURE, DANGEROUS_SITUATION, TRAFFIC_ACCIDENT, PARTICIPANT_BEHAVIOR, PARTICIPANT_HAZARD, WASTE_ILLEGAL_DUMPING, BIOLOGICAL_HAZARD, OTHER]
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /api/reports/report/{id}/status:
    patch:
      summary: Update report status
      tags: [reports]
      operationId: patchReportsReportStatus
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/X-Request-ID'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: status
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '404': { $ref: '#/components/responses/Error' }
  /api/reports/reports:
    get:
      summary: List reports (paginated)
      tags: [reports]
      security: [ { bearerAuth: [] } ]
      operationId: getReportsList
      parameters:
        - $ref: '#/components/parameters/X-Request-ID'
        - in: query
          name: page
          schema: { type: integer, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1 }
        - in: query
          name: sort
          schema: { type: string }
        - in: query
          name: direction
          schema: { type: string, enum: [asc, desc] }
      responses:
        '200':
          description: Reports page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '400': { $ref: '#/components/responses/Error' }
  /api/reports/report/{id}:
    get:
      summary: Get report by id
      tags: [reports]
      security: [ { bearerAuth: [] } ]
      operationId: getReportById
      parameters:
        - $ref: '#/components/parameters/X-Request-ID'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '404': { $ref: '#/components/responses/Error' }
  /api/reports/reports/verified:
    get:
      summary: List verified reports
      tags: [reports]
      security: [ { bearerAuth: [] } ]
      operationId: getReportsVerified
      parameters:
        - $ref: '#/components/parameters/X-Request-ID'
      responses:
        '200':
          description: Verified reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '400': { $ref: '#/components/responses/Error' }
...